Create main.py
NEW MONETARY ANALYSIS TOOL 
(THE DAILY LIQUIDITY DATASET)
Engin YILMAZ
The term of liquidity is complex and unclear concept in the economic literature. This paper tries to explain this concept in practice for Turkish economy. The statistics of EFT (Electronic Fund Transfer) system and Central Bank of Turkish Republic’s analytical balance sheet are used to explain what the liquidity is and how the liquidity situation changes in the Turkish financial system. The liquidity is classified as “market liquidity”. The EFT system is used to materialize the market liquidity and the statistics of EFT system are used to explain the volume of market liquidity. The statistics of Central Bank of Turkish Republic’s analytical balance sheet are utilized for illustrating the liquidity situation which is originated from market liquidity. The daily liquidity dataset for Central Bank of Turkish Republic is created and the determinants of the liquidity situation are analysed from perspective of analytical balance sheet. This study identifies that the increase in the currency issued and government sector deposits are main determinants of the liquidity deficit and also concludes that the liquidity surplus results from the increment in net foreign assets and domestic assets.